feat: default Tips on, remove the Send tab, and restrict Send Cash to tip chats
Enable Tips by default and remove the phone-contact Send experience. Send
Cash now only happens inside a (tip) conversation — the contact/phone
recipient surfaces are gone. Tip-specific entry points (scan card / tip
links) come in a follow-up PR.

Tips enabled by default:
- canUseTips returns true; the enableTips beta flag is removed; drop the
  tipcard beta opt-in in TipFlow.

Remove the Send tab + contact/phone send surfaces:
- Remove the Send button from the scan bottom bar.
- Delete the recipient picker (SendRootScreen/RecipientPickerScreen and its
  helpers) and the `.send` sheet/stack.
- Remove the flipcash://send deep link, the home-screen Send quick action,
  and the "Send Cash" Siri/App Intent (deletes the App Intents target).
- Remove the flipcash://chat/{phone} (.chatContact) deep link; chat deeplinks
  and pushes only open tip DMs now (contact DMs are no longer navigable).
- Remove the dmConversation destination and the ConversationContext.contact
  case; ConversationScreen now only renders tip DMs. Extract the shared row
  chrome (RecipientRowScaffold/RecipientRowAccessory) used by the Tips list.
- Remove Session.canSend and guard the /chat/{id}/send deeplink to tip targets.

In-conversation Send Cash (the button + the "Send Cash" push action) is
unchanged and stays labeled "Send Cash"; it now applies only to tip DMs.
SendTarget and SendAmountViewModel keep their (now-dormant) contact support.

Tests: delete the recipient-picker and send-stack suites; update Route,
QuickActions, and AddMoney routing tests; drop enableTips from the Tips UI
smoke overrides.
