+++ Flipcash/Core/Controllers/BetaFlags.swift
@@ -114,7 +114,6 @@ extension BetaFlags {
 
         case vibrateOnScan
         case enableCoinbase
-        case enableTips
         case enableBlocking
 
         var id: String {
@@ -127,8 +126,6 @@ extension BetaFlags {
                 return "Vibrate on scan"
             case .enableCoinbase:
                 return "Enable Coinbase"
-            case .enableTips:
-                return "Tips"
             case .enableBlocking:
                 return "Blocking"
             }
@@ -140,8 +137,6 @@ extension BetaFlags {
                 return "If enabled, the device will vibrate to indicate that the camera has registered the code on the bill"
             case .enableCoinbase:
                 return "If enabled, Coinbase onramp will be available regardless of region"
-            case .enableTips:
-                return "If enabled, the Tips tab is available from the scan screen"
             case .enableBlocking:
                 return "If enabled, you can block users from a tip chat and manage blocked users in My Account"
             }
@@ -152,7 +147,6 @@ extension BetaFlags {
             switch self {
             case .vibrateOnScan:   return .developer
             case .enableCoinbase:  return .developer
-            case .enableTips:      return .publicBeta
             case .enableBlocking:  return .developer
             }
         }
