+++ Flipcash/Core/Screens/Main/ScanBottomBar.swift
@@ -8,14 +8,11 @@ import FlipcashUI
 
 struct ScanBottomBar: View {
     let toast: String?
-    let showSend: Bool
-    let sendBadgeCount: Int
     let showTips: Bool
     let tipsBadgeCount: Int
     let onGive: () -> Void
     let onWallet: () -> Void
     let onDiscover: () -> Void
-    let onSend: () -> Void
     let onTips: () -> Void
 
     var body: some View {
@@ -33,16 +30,6 @@ struct ScanBottomBar: View {
                 action: onGive
             )
             .accessibilityIdentifier("scan-cash-button")
-
-            if showSend {
-                LargeButton(
-                    title: "Send",
-                    image: Image(.Icons.send),
-                    badgeCount: sendBadgeCount,
-                    action: onSend
-                )
-                .accessibilityIdentifier("scan-send-button")
-            }
 
             if showTips {
                 LargeButton(
