+++ Flipcash/Core/Screens/Main/Tips/RecipientRow.swift
@@ -1,0 +1,151 @@
+//
+//  RecipientRow.swift
+//  Flipcash
+//
+
+import SwiftUI
+import FlipcashCore
+import FlipcashUI
+
+/// Where a row's trailing accessory sits.
+enum RecipientRowAccessoryPlacement {
+    /// Centered in a full-height trailing column; the subtitle stays in the
+    /// column beside it.
+    case trailingColumn
+    /// On the title's line (Messages/Mail style); the subtitle spans the row's
+    /// full width below — the merged conversation feed.
+    case titleLine
+}
+
+/// The chrome every conversation row shares: a full-row button with avatar,
+/// title/subtitle, and an accessory. Used by the Tips conversations list.
+struct RecipientRowScaffold<Trailing: View>: View {
+
+    let avatarID: String
+    let title: String
+    let subtitle: String?
+    let imageData: Data?
+    var blurhash: String? = nil
+    var accessoryPlacement: RecipientRowAccessoryPlacement = .trailingColumn
+    let accessibilityLabel: String
+    let onTap: () -> Void
+    @ViewBuilder let trailing: Trailing
+
+    var body: some View {
+        Button(action: onTap) {
+            RecipientRowBody(
+                avatarID: avatarID,
+                title: title,
+                subtitle: subtitle,
+                imageData: imageData,
+                blurhash: blurhash,
+                accessoryPlacement: accessoryPlacement
+            ) {
+                trailing
+            }
+        }
+        .recipientRowChrome(accessibilityLabel: accessibilityLabel)
+    }
+}
+
+/// The visual content of a conversation row: avatar, title/subtitle, and an
+/// accessory placed per `accessoryPlacement`.
+struct RecipientRowBody<Trailing: View>: View {
+
+    let avatarID: String
+    let title: String
+    let subtitle: String?
+    let imageData: Data?
+    var blurhash: String? = nil
+    var accessoryPlacement: RecipientRowAccessoryPlacement = .trailingColumn
+    @ViewBuilder let trailing: Trailing
+
+    var body: some View {
+        HStack(spacing: 12) {
+            ContactAvatarView(
+                id: avatarID,
+                displayName: title,
+                imageData: imageData,
+                blurhash: blurhash,
+                size: 44
+            )
+            VStack(alignment: .leading, spacing: 2) {
+                HStack(alignment: .top, spacing: 6) {
+                    Text(title)
+                        .font(.appTextMedium)
+                        .foregroundStyle(Color.textMain)
+                        .lineLimit(1)
+                    if accessoryPlacement == .titleLine {
+                        Spacer(minLength: 12)
+                        trailing
+                    }
+                }
+                if let subtitle {
+                    Text(subtitle)
+                        .font(.appTextSmall)
+                        .foregroundStyle(Color.textSecondary)
+                        .lineLimit(1)
+                        .contentTransition(.opacity)
+                        .transition(.opacity)
+                }
+            }
+            // Cross-fade the subtitle when it changes — notably the "Typing…" ⇄ last-message swap, but
+            // also a fresh message preview replacing the previous one.
+            .animation(.easeInOut(duration: 0.2), value: subtitle)
+            if accessoryPlacement == .trailingColumn {
+                Spacer(minLength: 12)
+                trailing
+            }
+        }
+        .contentShape(Rectangle())
+    }
+}
+
+private extension View {
+    /// Row chrome shared by every conversation row: list insets, clear
+    /// background, separator tint, and single-element button accessibility.
+    func recipientRowChrome(accessibilityLabel: String) -> some View {
+        self
+            .buttonStyle(.plain)
+            .listRowInsets(EdgeInsets(top: 14, leading: 20, bottom: 14, trailing: 20))
+            .listRowBackground(Color.clear)
+            .listRowSeparatorTint(.rowSeparator)
+            .accessibilityElement(children: .ignore)
+            .accessibilityLabel(Text(accessibilityLabel))
+            .accessibilityAddTraits(.isButton)
+    }
+}
+
+/// A row's trailing accessory: a relative timestamp (or "Unknown Contact"
+/// pill), an unread dot, or a disclosure chevron.
+struct RecipientRowAccessory: View {
+
+    let timestamp: Date
+    let isUnknown: Bool
+    let hasUnread: Bool
+
+    var body: some View {
+        HStack(spacing: 6) {
+            if isUnknown {
+                Text("Unknown Contact")
+                    .fixedSize(horizontal: true, vertical: false)
+                    .pill()
+            } else {
+                Text(timestamp.formattedRelatively(useTimeForToday: true))
+                    .font(.appTextSmall)
+                    .foregroundStyle(hasUnread ? Color.unreadIndicator : Color.textSecondary)
+            }
+            if hasUnread {
+                Circle()
+                    .fill(Color.unreadIndicator)
+                    .frame(width: 9, height: 9)
+            } else {
+                Image(systemName: "chevron.right")
+                    .resizable()
+                    .aspectRatio(contentMode: .fit)
+                    .frame(width: 9, height: 12)
+                    .foregroundStyle(Color.textSecondary)
+            }
+        }
+    }
+}
