+++ FlipcashUITests/Smoke/BlockUnblockSmokeTests.swift
@@ -24,8 +24,8 @@ import XCTest
 final class BlockUnblockSmokeTests: BaseUITestCase {
 
     override var requiresAuthentication: Bool { true }
-    // Tips tab needs `enableTips`; the Block affordance needs `enableBlocking`.
-    override var enabledBetaFlags: [String] { ["enableTips", "enableBlocking"] }
+    // The Block affordance needs `enableBlocking`; Tips is always on now.
+    override var enabledBetaFlags: [String] { ["enableBlocking"] }
 
     /// The counterpart blocked during the test, cleared once unblocked. Non-nil
     /// in `tearDown` means the round-trip failed after the block — clean it up.
